// src/modules/marketplace/common/marketplace.enums.ts
export enum AvatarCategoryEnum {
  HAIRS = 'Hairs',
  TOPS = 'Tops',
  BOTTOMS = 'Bottoms',
}

export enum AvatarItemNameEnum {
  MALELONGHAIRS = 'MalelongHairs',
  MALEKOREANHAIRS = 'MalekoreanHairs',
  MALESHORTHAIRS = 'MaleShortHairs',
  FEMALELONGHAIRS = 'FeMalelongHairs',
  FEMALEMEDIUMHAIRS = 'FeMaleMediumHairs',
  FEMALESHORTHAIRS = 'FeMaleShortHairs',
  MALEHOOD = 'MaleHood',
  MALETSHIRT = 'Maletshirt',
  MALELONGSLEVESTSHIRT = 'Malelongslevestshirt',
  FEMALESSLEVESLESSTOP = 'FeMalessleveslesstop',
  FEMALESSLEVESSTOP = 'FeMalesslevesstop',
  FEMALECOAT = 'FeMalecoat',
  MALEUPPERSHORT = 'MaleupperShort',
  MALELOWERSHORT = 'Malelowershort',
  MALEPENT = 'Malepent',
}

export enum AvatarGenderEnum {
  MALE = 'Male',
  FEMALE = 'Female',
}

export enum AvatarColorEnum {
  BLACK = 'Black',
  BLUE = 'Blue',
  CYAN = 'Cyan',
  GREEN = 'Green',
  GREY = 'Grey',
  MAGENTA = 'Magenta',
  RED = 'Red',
  WHITE = 'White',
  YELLOW = 'Yellow',
}

export enum AvatarStatusEnum {
  ACTIVE = 'Active',
  DISABLED = 'Disabled',
}

export enum UserAvatarCategoryEnum {
  AVATAR = 'Avatar',
  HAIRS = 'Hairs',
  TOP = 'Top',
  BOTTOM = 'Bottom',
  SHOES = 'Shoes',
  SKINTONE = 'Skintone'
}